Menambah class Upload referensi
