+++ Tubes_KPL_RojakSquad/src/Referensi/Buku.java
@@ -20,20 +20,9 @@ public class Buku extends referensi {
         return this.publisher;
     }
 
-    @Override
-    public void uploadReferensi(String tipe){
 
-        List listR = this.getList();
-        listR.add(this);
-        Map dataR = this.getDB();
-        dataR.put(tipe,listR);
+  
 
-    }
 
-    @Override
-    public void info(){
-       
-
-    }
 
 }
