+++ Tubes_KPL_RojakSquad/src/Referensi/referensi.java
@@ -14,9 +14,6 @@ public class referensi {
 
     enum Jenis{Internasional, Nasional};
 
-    private HashMap<String,List> dataBase = new HashMap<String,List>();
-    private List<referensi> listR = new ArrayList<referensi>();
-
     public referensi(String judul, String penulis, String topik, String jenis) throws IllegalArgumentException{
         if (judul == null || judul.isEmpty()){
             throw new IllegalArgumentException("Judul referensi kosong");
@@ -35,6 +32,8 @@ public class referensi {
         this.penulis = penulis;
         this.topik = topik;
         this.jenis = jenis; 
+
+        
         this.rating = 0;
         
     }
@@ -76,46 +75,7 @@ public class referensi {
         return this.uploadDate; 
 	}
 
-    public HashMap getDB(){
-        return this.dataBase;
-    }
 
-    public List getList(){
-        return this.listR;
-    }
-
-    public void uploadReferensi(String tipe){
-
-        
-        
-        List listR = this.getList();
-        listR.add(this);
-        Map dataR = this.getDB();
-        dataR.put(tipe,listR);
-
-    }
-
-
-    public void info(){
-
-        for (Map.Entry<String,List> entry: this.dataBase.entrySet()) {
-         
-            System.out.println("Key:" +entry.getKey());
-
-            for(int i=0;i<entry.getValue().size(); i++) {
-                System.out.println("------------------");
-                referensi r1 = (referensi)entry.getValue().get(i);
-                System.out.println("Judul:" +r1.getJudul());
-                System.out.println("Penulis:" + r1.getPenulis());
-                System.out.println("Topik:" + r1.getTopik());
-                System.out.println("Jenis:" + r1.getJenis());
-                System.out.println("Rating:" + r1.getRating());
-                System.out.println("Tanggal Upload:" + r1.getDates());
-                System.out.println("------------------");
-            }
-        }
-
-    }
